refactor: :recycle: update helpers folder name to utils and rename files
