+++ src/utils/spotlight.ts
